requirements.txt
gradio
ollama
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib